# api/openapi.yaml
openapi: 3.0.1
info:
  title: volunteer-portal-api
  version: 1.30.0
  contact:
    email: dev@russian.rs
    url: https://russian.rs
servers:
  - url: https://REDACTED_HOST/api
    description: Test Environment
  - url: https://portal.russian.rs/api
    description: Prod Environment
paths:
  /user/account:
    get:
      operationId: getCurrentAccount
      tags:
        - user
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoDto"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      operationId: createUser
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoDto"
  /user/account/{id}/program/{code}:
    patch:
      operationId: setProgram
      tags:
        - user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            nullable: false
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoDto"
  /user/account/{id}/project/{code}:
    patch:
      operationId: setProject
      tags:
        - user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            nullable: false
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoDto"
  /user/account/{id}/deactivate:
    post:
      operationId: deactivateAccount
      tags:
        - user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            nullable: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoDto"
  /user/account/{id}/activate:
    post:
      operationId: activateAccount
      tags:
        - user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            nullable: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoDto"
  /user/account/{id}/contracts:
    post:
      operationId: updateContracts
      description: Обновить список договоров волонтера
      tags:
        - user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            nullable: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/ContractDto"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoDto"
  /user/account/{id}/residence-permits:
    put:
      operationId: updateResidencePermits
      description: Обновить список ВНЖ волонтера
      tags:
        - user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/ResidencePermitDto"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoDto"
  /user/account/{id}/residence-permits/{permitId}:
    delete:
      operationId: deleteResidencePermit
      description: Удалить ВНЖ волонтера (только для администраторов)
      tags: [ user ]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
        - { in: path, name: permitId, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoDto"
  /user/info/{login}:
    get:
      operationId: getInfo
      tags:
        - user
      parameters:
        - name: login
          in: path
          description: User login
          required: true
          schema:
            type: string
            minLength: 3
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoDto"
        "500":
          $ref: "#/components/responses/ServerError"

    patch:
      operationId: updateInfo
      tags:
        - user
      parameters:
        - name: login
          in: path
          description: User login
          required: true
          schema:
            type: string
            minLength: 3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInfoUpdateRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoDto"
        "500":
          $ref: "#/components/responses/ServerError"
  /user/search:
    post:
      operationId: searchUsers
      tags:
        - user
      parameters:
        - name: searchQuery
          in: query
          required: true
          schema:
            type: string
        - name: pageRequest
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/PageRequest"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSearchFilter"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPageResponse"
...
  /user/avatar:
    post:
      operationId: setAvatar
      tags:
        - user
      parameters:
        - name: avatarId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoDto"
  /user/resolve:
    post:
      operationId: resolveUsers
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              default: [ ]
              items:
                type: string
                maxLength: 100
                minLength: 1
                nullable: false
                description: username
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                default: [ ]
                items:
                  $ref: "#/components/schemas/UserInfoDto"
...
components:
  responses:
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    PageRequest:
      type: object
      description: Paging object
      properties:
        pageNumber:
          type: integer
          format: int32
          default: 0
          nullable: false
        pageSize:
          type: integer
          format: int32
          default: 10
          nullable: false
        sort:
          type: array
          default: [ ]
          items:
            type: string
    PageResponse:
      type: object
      nullable: false
      description: Page info
      required:
        - page
        - size
        - totalPages
        - totalElements
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
    ErrorResponse:
      type: object
      description: Общий объект для сообщений об ошибках
      required:
        - message
      properties:
        message:
          type: string
          description: Описание ошибки
          nullable: false
    GenderEnumDto:
      type: string
      enum:
        - MALE
        - FEMALE
    UserCreateRequest:
      type: object
      required:
        - email
        - username
        - fullName
        - contract
      properties:
        email:
          type: string
          format: email
          maxLength: 100
        username:
          type: string
          minLength: 2
          maxLength: 100
        fullName:
          type: string
          minLength: 2
          maxLength: 100
        contract:
          $ref: "#/components/schemas/ContractDto"
    UserInfoDto:
      type: object
      description: Дополнительная информация о пользователе
      required:
        - id
        - username
        - email
        - fullName
        - groups
        - active
      properties:
        id:
          type: integer
          nullable: false
        username:
          type: string
          nullable: false
        email:
          type: string
          nullable: false
        fullName:
          type: string
          nullable: false
        city:
          type: string
        postalCode:
          type: string
        address:
          type: string
        birthDate:
          type: string
          format: date
        telegram:
          type: string
        phone:
          type: string
        program:
          $ref: "#/components/schemas/ProgramDto"
        project:
          $ref: "#/components/schemas/ProjectDto"
        avatar:
          $ref: "#/components/schemas/FileInfoDto"
        groups:
          type: array
          uniqueItems: true
          items:
            type: string
        active:
          type: boolean
          nullable: false
        gender:
          $ref: "#/components/schemas/GenderEnumDto"
        contracts:
          type: array
          default: [ ]
          items:
            $ref: "#/components/schemas/ContractDto"
        residencePermits:
          type: array
          default: [ ]
          items:
            $ref: "#/components/schemas/ResidencePermitDto"
    UserInfoUpdateRequest:
      type: object
      description: обновить информацию о пользователе
      properties:
        city:
          type: string
        postalCode:
          type: string
        address:
          type: string
        birthDate:
          type: string
          format: date
        telegram:
          type: string
        phone:
          type: string
        gender:
          $ref: "#/components/schemas/GenderEnumDto"
      additionalProperties: false
    UserPageResponse:
      type: object
      nullable: false
      required:
        - page
        - content
      properties:
        page:
          $ref: "#/components/schemas/PageResponse"
        content:
          type: array
          default: [ ]
          nullable: false
          items:
            $ref: "#/components/schemas/UserInfoDto"
    FileInfoDto:
      type: object
      nullable: true
      required:
        - id
        - name
        - link
        - size
      properties:
        id:
          type: string
          nullable: false
        name:
          type: string
          nullable: false
        link:
          type: string
          nullable: false
        contentLength:
          type: integer
          format: int64
          nullable: false
          description: Size (bytes)
...
    ContractDto:
      type: object
      required:
        - id
        - startDate
        - endDate
        - type
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        startDate:
          type: string
          format: date
          nullable: false
        endDate:
          type: string
          format: date
          nullable: false
        type:
          $ref: "#/components/schemas/ContractTypeEnum"
    ResidencePermitDto:
      type: object
      required:
        - id
        - nationality
        - registrationNumber
        - validUntil
        - purposeOfStay
        - issuingAuthority
        - stateOfBirth
        - issuingDate
        - identityNumber
        - frontSidePhoto
        - backSidePhoto
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        nationality:
          type: string
          nullable: false
        registrationNumber:
          type: string
          nullable: false
        validUntil:
          type: string
          format: date
          nullable: false
        purposeOfStay:
          type: string
          nullable: false
        note:
          type: string
          nullable: true
        issuingAuthority:
          type: string
          nullable: false
        stateOfBirth:
          type: string
          nullable: false
        issuingDate:
          type: string
          format: date
          nullable: false
        identityNumber:
          type: string
          nullable: false
        frontSidePhoto:
          $ref: "#/components/schemas/FileInfoDto"
        backSidePhoto:
          $ref: "#/components/schemas/FileInfoDto"
...
    UserSearchFilter:
      type: object
      nullable: false
      properties:
        onlyActive:
          type: boolean
          nullable: true
          description: Показывать только активных пользователей
        onlyInactive:
          type: boolean
          nullable: true
          description: Показывать только неактивных пользователей (onlyActive приоритетней если указаны оба)
        program:
          type: string
          nullable: true
          description: Показывать только пользователей по указанной программе, пустая строка значит показать пользователей без программы
          example: "IT"
        project:
          type: string
          nullable: true
          description: Показывать только пользователей по указанному проекту, пустая строка значит показать пользователей без проекта
          example: "DESIGN_3D"
    ContractTypeEnum:
      type: string
      enum:
        - ASSOCIATED
        - REGULAR
...
    ProgramDto:
      type: object
      required:
        - code
        - nameRu
        - nameEn
        - nameSr
      properties:
        code:
          type: string
        nameRu:
          type: string
        nameEn:
          type: string
        nameSr:
          type: string
        projectCodes:
          type: array
          description: List of project codes related to this program
          items: { type: string }
    ProjectDto:
      type: object
      required:
        - code
        - nameRu
        - nameEn
        - nameSr
      properties:
        code:
          type: string
        nameRu:
          type: string
        nameEn:
          type: string
        nameSr:
          type: string
        programCode:
          type: string
          description: Owning program code
...